Add Home and NotFound pages
