+++ src/pages/NotFound/NotFound.scss
@@ -1,0 +1,19 @@
+@import '../../styles/_colors';
+
+.not-found {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+
+  a {
+    color: $red;
+    text-decoration: none;
+
+    &:hover {
+      color: $dark-red;
+      text-decoration: underline;
+    }
+  }
+}
