+++ webpack.dev.js
@@ -64,5 +64,6 @@ module.exports = {
     compress: true,
     open: true,
     hot: true,
+    historyApiFallback: true,
   },
 };
